backend/requirements.txt
flask
transformers
torch
torchvision
pillow
flask_cors